Refactored handle_privmsg so all joinmarket logic in on_privmsg in message_channel, and removed unused self.waiting module var. Also moved check_for_orders to message_channel.
